update user auth by addind password hashing
