+++ frontend/login.html
@@ -54,7 +54,7 @@
               },
               body: JSON.stringify({
                 email: email.value,
-                password_hash: password.value
+                password: password.value
               })
             }).then(response =>
               response.json()
