+++ frontend/register.html
@@ -66,7 +66,7 @@
               last_name: last_name.value,
               phone_number: phone_number.value,
               email: email.value,
-              password_hash: password.value,
+              password: password.value,
             })
           }).then(response => {
             console.log(response)
